change in mongo sample project api out put

+++ Samples/SampleWebApiApplicationWithMongoDb/Controllers/PersonController.cs
@@ -1,11 +1,13 @@
 ﻿using Hephaestus.Repository.Abstraction.Contract;
+using Hephaestus.Repository.Abstraction.Exceptions;
+using Mapster;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using SampleWebApiApplicationWithMongoDb.Models;
 using SampleWebApiApplicationWithMongoDb.Persistence;
+using SampleWebApiApplicationWithMongoDb.ViewModels;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -26,50 +28,57 @@ namespace SampleWebApiApplicationWithMongoDb.Controllers
         }
 
         [HttpPost]
-        public async Task<ActionResult> Create([FromBody] CreatePersonRequest request)
+        public async Task<ActionResult<PersonViewModel>> Create([FromBody] CreatePersonRequest request)
         {
             var person = PersonEntity.Create(request.FirstName, request.LastName);
             await _personRepository.AddAsync(person, CancellationToken.None);
             await _unitOfWork.CommitAsync();
 
-            return Created($"/Person/{person.Id}", person);
+            return Created($"/Person/{person.Id}", person.Adapt<PersonViewModel>());
         }
 
         [HttpPut]
-        public async Task<ActionResult> Update([FromQuery] Guid id, [FromBody] CreatePersonRequest request)
+        public async Task<ActionResult<PersonViewModel>> Update([FromQuery] Guid id, [FromBody] CreatePersonRequest request)
         {
             var person = await _personRepository.GetAsync(id, CancellationToken.None);
+            if (person == null)
+                throw new EntityNotFoundException($"Unable to find Person with ID '{id}'");
+
             person.Update(request.FirstName, request.LastName);
             await _personRepository.UpdateAsync(person, CancellationToken.None);
             await _unitOfWork.CommitAsync();
 
-            return Created($"/Person/{person.Id}", person);
+            return Created($"/Person/{person.Id}", person.Adapt<PersonViewModel>());
         }
 
         [HttpDelete("/id")]
         public async Task<ActionResult> Delete([FromQuery] Guid id)
         {
             var person = await _personRepository.GetAsync(id, CancellationToken.None);
+            if (person == null)
+                throw new EntityNotFoundException($"Unable to find Person with ID '{id}'");
+
+            person.Delete();
             await _personRepository.DeleteAsync(person, CancellationToken.None);
             await _unitOfWork.CommitAsync();
 
-            return Created($"/Person/{person.Id}", person);
+            return NoContent();
         }
 
         [HttpGet("/id")]
-        public async Task<PersonEntity> Get([FromQuery] Guid id)
+        public async Task<PersonViewModel> Get([FromQuery] Guid id)
         {
             var person = await _personRepository.GetAsync(id, CancellationToken.None);
 
-            return person;
+            return person.Adapt<PersonViewModel>();
         }
 
         [HttpGet()]
-        public async Task<IEnumerable<PersonEntity>> GetList()
+        public async Task<IEnumerable<PersonViewModel>> GetList()
         {
             var persons = await _personRepository.GetListAsync(CancellationToken.None);
 
-            return persons;
+            return persons.Adapt<IEnumerable<PersonViewModel>>();
         }
     }
 }
